==> 30 cali_plots.Rmd <==
---
title: "California EJ Scoping"
author: "Will Wheeler"
date: "2/3/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

library(tidyverse)
library(sf)


theme_set(theme_minimal() + 
            theme(plot.title.position = "plot", 
                  plot.caption.position = "plot"))
```

This is an initial attempt to demonstrate the feasibility of combining different sources of data on drinking water systems and their customers to provide a descriptive analysis of EJ as it applies to drinking water. To begin, we focus on California, which has the best available data. 

## Data Sources

We use a number of data sources, including: 

- Geospatial data on water system boundaries 
- Chemical sampling data for drinking water systems obtained from the California Water Board (https://www.waterboards.ca.gov/drinking_water/certlic/drinkingwater/EDTlibrary.html)
- Electronic Annual Report (EAR) data on water systems, again from the Board. These reports contain detailed information on systems, including on water rates.
- Sociodemographic information from the American Community Survey (ACS)

All of these data are matched to systems using the boundary data and interpolated, where necessary, using the areal package in R. 

```{r load_data}

load(here::here("data", "cali_interpolated_all.Rda"))

```


```{r bill_by_income, echo=FALSE}

## graph bill by income 
cali_graph1 <- cali_interpolated_all %>%
  select(-geometry) %>%
  filter(!is.na(median_income), !is.na(wr_12_hcf_total_w_bill),
         wr_12_hcf_total_w_bill > 0) %>%
  mutate(pct_income = wr_12_hcf_total_w_bill*12/median_income)

```


We begin with a number of graphs showing how water bills and sociodemographics interact. We use a self-reported normalized water bill from the EAR survey data. We use the normalized bill for 12 hundred cubic feet (about 9000 gallons) to represent an approximate bill for a family.

First, the normalized bill by median income for the system. 

```{r bill_by_MHI}

ggplot(data = cali_graph1 %>% 
         filter(wr_12_hcf_total_w_bill < 1000, wr_12_hcf_total_w_bill > 0), 
       aes(x = median_income, y = wr_12_hcf_total_w_bill)) +
  labs(title = " Normalized Water Bill by Median Income") +
  xlab("Normalized Bill, $/month") +
  geom_point() 

```

And the percent of median income by median income. 

```{r bill_by_MHI_trim}

ggplot(data = cali_interpolated_all %>% 
         filter(pct_income < 0.5), 
       aes(x = median_income, y = pct_income)) +
  labs(title = "Median Income by Normalized Bill as Percent Income") +
  geom_point() +
  labs(caption = "One system dropped, pct income > 2") 

```
Bill expressed as percent of income by percent black and percent Hispanic.

```{r pct_black_by_pct_income}

ggplot(data = cali_interpolated_all %>% 
         filter(pct_income < 0.5), 
       aes(x = pct_black, y = pct_income)) +
  labs(title = "Percent Black by Pct Income of Normalized Water Bill") +
  geom_point() +
  labs(caption = "One system dropped, pct income > 2") 

```


```{r pct_hispanic_by_pct_income}


ggplot(data = cali_interpolated_all %>% 
         filter(pct_income < 0.5), 
       aes(x = pct_hispanic, y = pct_income)) +
  labs(title = "Percent Hispanic by Pct Income of Normalized Water Bill") +
  geom_point() 


```





Now compare violations (since 2018) to demographics. I did a quick summary and this is almost all MCL violationss, so I didn't clean the violation data.

```{r violations_by_income}

ggplot(data = cali_interpolated_all %>% 
         mutate(viols_since_2018 = replace_na(viols_since_2018, 0)), 
       aes(x = median_income, y = viols_since_2018)) +
  geom_point() + 
  #geom_smooth(method = "loess", se = FALSE) +
  labs(title = "Number of Violations by Median Income",
       x = "System MHI", y = "Number of Violations")

ggsave("violations_by_income.png", height = 5, width = 9, units = "in")
```



```{r viols_by_pct_black}

ggplot(data = cali_interpolated_all, 
       aes(x = pct_black, y = viols_since_2018)) +
  geom_point() +
  labs(title = "Number of Violations by Pct Black")

```



```{r viols_by_pct_hispanic}



ggplot(data = cali_interpolated_all, 
       aes(x = pct_hispanic, y = viols_since_2018)) +
  geom_point() +
  labs(title = "Number of Violations by Percent Hispanic") +
  xlab("Percent Hispanic") + ylab("Mean Nitrate, mg/l")

ggsave("viols_by_pct_hispanic.png")

```





```{r nitrate_by_pct_black}

ggplot(data = cali_interpolated_all, 
       aes(x = pct_black, y = mean_nitrate)) +
  geom_point() +
  labs(title = "Mean Nitrate Levels by Pct Black", subtitle = "2010-2015")

```



```{r nitate_by_pct_hispanic}



ggplot(data = cali_interpolated_all, 
       aes(x = pct_hispanic, y = mean_nitrate)) +
  geom_point() +
  #geom_smooth(method = "loess", se = FALSE) +
  labs(title = "Mean Nitrate Levels by Percent Hispanic", subtitle = "California, 2010-2015") +
  xlab("Percent Hispanic") + ylab("Mean Nitrate, mg/l")

ggsave("nitrate_hispanic_cali.png", height = 5, width = 9, units = "in")

```


<!--  ```{r thm_by_pct_black}  -->
<!--  ggplot(data = cali_interpolated_all,   -->
<!--         aes(x = pct_black, y = mean_thm)) + -->
<!--    geom_point() + -->
<!--    labs(title = "Mean THM Levels by Pct Black") -->
<!-- ``` -->


<!--  ```{r thm_by_pct_hispanic} -->

<!--  ggplot(data = cali_interpolated_all,  -->
<!--         aes(x = pct_hispanic, y = mean_thm)) + -->
<!--    geom_point() + -->
<!--    labs(title = "Mean THM Levels by Pct Hispanic") -->
<!--   ``` -->

To-dos: 

- Water system boundaries are available for four other states, how feasible is extending analysis (unclear what rate data is available)
- Currently using ACS block group data, is it possible to use a finer geographic distinction?
- Using reported standardized water rates (12 HCF), compare with calculated rates (also some standardized rates are missing)
- Look at other charges as well
- Much richer data on shutoffs in EARs
- Evaluate quality/readability of CCRs
- Add coliform data
- Percent White variable is clearly wrong
- Verify that we are using correct sample point codes
- Self-intersecting polygons in system map
- Not all systems have interpolated data
- Investigate/compare different interpolation methods
